Update README.md
Task: NUMBER GUESSING GAME
It represents a simple guessing the number game that allows the user to control how many attempts the player has to guess the number. The game informs the player whether their guess is too low or too high compared to the target number. After the game is over, it provides result on whether the player guessed the number or ran out of attempts.
Thank you .